repo1

+++ TransportPass/src/domain/Bilet.java
@@ -6,11 +6,11 @@ import java.util.Objects;
 public class Bilet extends Entity<Long> {
     private LocalDateTime dataIncepere;
     private LocalDateTime dataExpirare;
-    private Long pret;
+    private Double pret;
     private String tip;
     private Client codClient;
 
-    public Bilet(Long id,LocalDateTime dataIncepere, LocalDateTime dataExpirare, Long pret, String tip, Client codClient) {
+    public Bilet(Long id,LocalDateTime dataIncepere, LocalDateTime dataExpirare, Double pret, String tip, Client codClient) {
         this.id = id;
         this.dataIncepere = dataIncepere;
         this.dataExpirare = dataExpirare;
@@ -43,11 +43,11 @@ public class Bilet extends Entity<Long> {
         this.dataExpirare = dataExpirare;
     }
 
-    public Long getPret() {
+    public Double getPret() {
         return pret;
     }
 
-    public void setPret(Long pret) {
+    public void setPret(Double pret) {
         this.pret = pret;
     }
